Bug fixes and performance and UI improvements
Attachment, settings, timetable widget, home overview items, bottom navigation and other UI improvements
many bug fixes - bottom navigation, homework, refreshing with 'failed to load' message,...
temporally implemented custom server loading because of holiday
And many more small bug fixes and performance and UI improvements

+++ app/src/main/java/cz/lastaapps/bakalariextension/api/homework/HomeworkParser.kt
@@ -98,7 +98,7 @@ class HomeworkParser {
                         safeJson(json, "Id"),
                         safeJson(json, "Name"),
                         safeJson(json, "Type"),
-                        json.getInt("Size")
+                        json.getLong("Size")
                     )
                 )
             }
